+++ src/commonMain/kotlin/kotools/types/text/NotBlankString.kt
@@ -24,7 +24,7 @@ import kotlin.jvm.JvmInline
 @SinceKotoolsTypes("4.0")
 public value class NotBlankString private constructor(
     private val value: String
-) {
+) : Comparable<NotBlankString> {
     internal companion object {
         infix fun of(value: String): Result<NotBlankString> = value
             .takeIf(String::isNotBlank)
@@ -36,6 +36,16 @@ public value class NotBlankString private constructor(
     public val length: StrictlyPositiveInt
         get() = value.length.toStrictlyPositiveInt()
             .getOrThrow()
+
+    /**
+     * Compares this string alphabetically with the [other] one for order.
+     * Returns zero if this string equals the [other] one, a negative number if
+     * it's less than the [other] one, or a positive number if it's greater than
+     * the [other] one.
+     */
+    @SinceKotoolsTypes("4.1")
+    override infix fun compareTo(other: NotBlankString): Int =
+        "$this".compareTo("$other")
 
     /** Returns this string as a [String]. */
     override fun toString(): String = value